separate types and api calls consistantly and remove duplicates
